requirements.txt
uvicorn
fastapi
pymysql
python-dotenv
sqlalchemy
pydentic